Simple hack to display multiple instances
Use the "x" key to show more copies, "z" key to show fewer. It's limited to 256 at the moment as I find there are artifacts with more instances. This is hacky and pretty much made just for demo purposes, so not sure whether I want to check it in just yet.