Replace keyframe with kurbo for tweening because of multiple reasons (#80)
There's an issue in keyframe where it produces garbage output and also
doesn't tween correctly.

Given the following input

```rust
keyframe::ease(
            keyframe::functions::BezierCurve::from(
                keyframe::mint::Vector2::from_slice(&[1.0, 0.0]),
                keyframe::mint::Vector2::from_slice(&[0.0, 1.0]),
            ),
            16.,
            64.,
            0.667,
        );
```
        
It produces `-8046854384`.

This is how it looks currently with keyframe, notice the small jitter
when the giraffe reaches the top and bottom:

[Screencast from 2025-11-14
08-24-32.webm](https://github.com/user-attachments/assets/df9c2dbf-6f0e-4be1-bcaf-48a1e7930310)

This is how a tween should look, smooth, as it should be represented,
https://lottiefiles.com/free-animation/meditating-giraffe-IAYfWtLb6E:

[Screencast from 2025-11-14
08-23-03.webm](https://github.com/user-attachments/assets/9d2dcf29-d4f5-4427-a3e9-f3383fd47950)