Actually tween with a cubic bezier curve (#81)
I totally misunderstood eval in the previous pr #80, sorry, I think this
is kosher but I would love someone else's eyes on this that are a bit
more experienced with linear algebra.

I dug into the kurbo code half the day and found `cubic_bez_coefs` and
`parameters` to get the parameters for `solve_cubic`.

It's looks correct in practice but not entirely sure if it's good in the
details.

Lottielab:

[Screencast from 2025-11-17
15-51-35.webm](https://github.com/user-attachments/assets/5022a628-8e0c-44e0-bae9-a83b91e7bf69)

Velato:

[Screencast from 2025-11-17
15-51-20.webm](https://github.com/user-attachments/assets/912a586e-b0c4-4dfb-ad88-348cf191e83d)