Just grab the x and y coordinates of the easing handles (#85)
Lottie supports an arbitrary amount of values in the spec for it to be
flexible.

Some lotties, like
[https://googlefonts.github.io/noto-emoji-animation/?icon.query=uni](https://googlefonts.github.io/noto-emoji-animation/?icon.query=uni)
has three or more components instead of just two to represent an easing
handle.

In our binary search
https://github.com/linebender/velato/blob/main/src/runtime/model/value.rs#L120,
we **probably** grab the first easing handle for the frame and the last
easing handle for the frame.

So in the unicorn case, we get the x and z components for each handle
instead of x and y; Which means the unicorn's stars are scaled very
weirdly using the z axis instead of y axis.

```json
{
  "o": {
    "x": [0.167, 0.167, 0.167],
    "y": [0.167, 0.167, 14.319]
  },
}
```

**See the stars around the unicorn**

Before

[Screencast from 2025-11-27
22-06-33.webm](https://github.com/user-attachments/assets/7e7277bb-6140-4f5e-8a8d-969c478a93bc)

After

[Screencast from 2025-11-27
22-06-11.webm](https://github.com/user-attachments/assets/ab23f929-7ce3-4445-975f-2119705d8ca7)